+++ index.html
@@ -226,7 +226,7 @@
       </div>
     </section>
   </template>
-  <script src="js/main.js"></script>
+  <script type = "module" src="js/main.js"></script>
   <!-- Экран загрузки изображения -->
   <template id="messages">
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
