+++ js/constant.js
@@ -1,0 +1,30 @@
+const NAME = [
+  'Акакий Георгиевич',
+  'Святослав Иванович',
+  'Ярослава Олеговна',
+  'Ольга Игоревна',
+  'Иван Васильевич',
+  'София Яновна',
+  'Александр Михайлович',
+  'Константин Данилович',
+];
+
+const DESCRIPTION = [
+  'котики отдыхают',
+  'я слушаю любимую музыку',
+  'кот смешно зевает',
+  'собака увидела первый снег',
+  'брат подскользнулся на люду',
+  'я ем',
+];
+
+const MESSAGE = [
+  'окей',
+  'привет',
+  'как твои дела?',
+  'чем занимаешься?',
+  'высыпаешься?',
+  'а я?',
+];
+
+export {NAME, MESSAGE, DESCRIPTION};
